third_party/blink/web_tests/fast/dom/MutationObserver/mutation-observer-constructor-expected.txt

Test the constructor of WebKitMutationObserver

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.WebKitMutationObserver is non-null.
PASS typeof WebKitMutationObserver.prototype.observe is "function"
PASS typeof WebKitMutationObserver.prototype.disconnect is "function"
PASS typeof observer.observe is "function"
PASS typeof observer.disconnect is "function"
PASS new MutationObserver({ handleEvent: function() {} }) threw exception TypeError: Failed to construct 'MutationObserver': parameter 1 is not of type 'Function'..
PASS new MutationObserver({}) threw exception TypeError: Failed to construct 'MutationObserver': parameter 1 is not of type 'Function'..
PASS new MutationObserver(42) threw exception TypeError: Failed to construct 'MutationObserver': parameter 1 is not of type 'Function'..
PASS new MutationObserver("foo") threw exception TypeError: Failed to construct 'MutationObserver': parameter 1 is not of type 'Function'..
PASS successfullyParsed is true

TEST COMPLETE
